ISSUE #3577 serialiseTicket method moved from utils/helper/ticket to schema/ticket test adjustments small chages in chat events doc MODEL_NEW_TICKET and MODEL_UPDATE_TICKET renamed
