+++ backend/tests/v5/unit/services/eventsListener/eventsListener.test.js
@@ -394,7 +394,7 @@ const testModelEventsListener = () => {
 		});
 
 		const updateTicketTest = async (isFederation) => {
-			const waitOnEvent = eventTriggeredPromise(events.MODEL_UPDATE_TICKET);
+			const waitOnEvent = eventTriggeredPromise(events.UPDATE_TICKET);
 			const data = {
 				teamspace: generateRandomString(),
 				project: generateRandomString(),
@@ -413,7 +413,7 @@ const testModelEventsListener = () => {
 			};
 
 			const event = isFederation ? chatEvents.FEDERATION_UPDATE_TICKET : chatEvents.CONTAINER_UPDATE_TICKET;
-			EventsManager.publish(events.MODEL_UPDATE_TICKET, data);
+			EventsManager.publish(events.UPDATE_TICKET, data);
 
 			await waitOnEvent;
 			expect(TicketLogs.addTicketLog).toHaveBeenCalledTimes(1);
@@ -430,17 +430,17 @@ const testModelEventsListener = () => {
 		};
 
 		test(`Should trigger addTicketLog and create a ${chatEvents.CONTAINER_UPDATE_TICKET} if there 
-				is a ${events.MODEL_UPDATE_TICKET} (Container)`, async () => {
+				is a ${events.UPDATE_TICKET} (Container)`, async () => {
 			await updateTicketTest(false);
 		});
 
 		test(`Should trigger addTicketLog and create a ${chatEvents.FEDERATION_UPDATE_TICKET} if there 
-				is a ${events.MODEL_UPDATE_TICKET} (Federation)`, async () => {
+				is a ${events.UPDATE_TICKET} (Federation)`, async () => {
 			await updateTicketTest(true);
 		});
 
 		const addTicketTest = async (isFederation) => {
-			const waitOnEvent = eventTriggeredPromise(events.MODEL_NEW_TICKET);
+			const waitOnEvent = eventTriggeredPromise(events.NEW_TICKET);
 			const data = {
 				teamspace: generateRandomString(),
 				project: generateRandomString(),
@@ -452,7 +452,7 @@ const testModelEventsListener = () => {
 			};
 
 			const event = isFederation ? chatEvents.FEDERATION_NEW_TICKET : chatEvents.CONTAINER_NEW_TICKET;
-			EventsManager.publish(events.MODEL_NEW_TICKET, data);
+			EventsManager.publish(events.NEW_TICKET, data);
 
 			await waitOnEvent;
 			expect(ChatService.createModelMessage).toHaveBeenCalledTimes(1);
@@ -466,12 +466,12 @@ const testModelEventsListener = () => {
 		};
 
 		test(`Should create a ${chatEvents.CONTAINER_NEW_TICKET} if there 
-				is a ${events.MODEL_NEW_TICKET} (Container)`, async () => {
+				is a ${events.NEW_TICKET} (Container)`, async () => {
 			await addTicketTest(false);
 		});
 
 		test(`Should create a ${chatEvents.FEDERATION_NEW_TICKET} if there 
-				is a ${events.MODEL_NEW_TICKET} (Federation)`, async () => {
+				is a ${events.NEW_TICKET} (Federation)`, async () => {
 			await addTicketTest(true);
 		});
 	});
